feat: new searchfield
BREAKING CHANGE: search-bar is no longer
