--- packages/core/src/components/search-bar.js
@@ -1,5 +1,0 @@
-module.exports = SearchBar = () => ({
-  '.search-bar': {
-    '@apply w-full relative': {},
-  },
-});
