+++ packages/core/src/components/search-field.js
@@ -1,0 +1,20 @@
+module.exports = SearchField = () => ({
+  '.sk-search-field': {
+    '&-md': {
+      '.sk-search-field-icon': {
+        '@apply w-[2rem] h-[2rem]': {},
+        svg: {
+          '@apply w-[2rem] h-[2rem]': {},
+        },
+      },
+    },
+    '&-lg': {
+      '.sk-search-field-icon': {
+        '@apply w-[2.2rem] h-[2.2rem]': {},
+        svg: {
+          '@apply w-[2.2rem] h-[2.2rem]': {},
+        },
+      },
+    },
+  },
+});
