--- packages/search-bar/globals.d.ts
@@ -1,1 +1,0 @@
-declare module '*.png';
