--- packages/search-bar/src/index.ts
@@ -1,1 +1,0 @@
-export * from './search-bar';
