+++ packages/searchfield/README.md
@@ -1,0 +1,5 @@
+## Installation
+
+```sh
+yarn add @sk-web-gui/searchfield
+```
