+++ packages/searchfield/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@sk-web-gui/search-bar",
-  "version": "0.1.20",
+  "name": "@sk-web-gui/searchfield",
+  "version": "0.1.0",
   "sideEffects": false,
   "license": "MIT",
   "main": "dist/cjs/index.js",
@@ -40,7 +40,8 @@
     "@mui/material": "^5.14.12",
     "@sk-web-gui/button": "0.1.22",
     "@sk-web-gui/forms": "0.1.37",
-    "@sk-web-gui/utils": "0.1.8"
+    "@sk-web-gui/utils": "0.1.8",
+    "@sk-web-gui/icon": "0.1.15"
   },
   "devDependencies": {
     "react": "^18.2.0"
